+++ src/test/unit/java/com/sdl/selenium/web/form/TextFieldTest.java
@@ -12,6 +12,10 @@ public class TextFieldTest {
                 {new TextField(), "//input"},
                 {new TextField("ID"), "//input[@id='ID']"},
                 {new TextField("ID").setType("text"), "//input[@id='ID' and @type='text']"},
+                {new TextField("ID").setPlaceholder("Search"), "//input[@id='ID' and @placeholder='Search']"},
+                {new TextField("ID").setPlaceholder(null), "//input[@id='ID']"},
+                {new TextField("ID").setPlaceholder("Search").setPlaceholder(null), "//input[@id='ID']"},
+                {new TextField("ID").setPlaceholder("Search").setClasses("cls"), "//input[@id='ID' and contains(concat(' ', @class, ' '), ' cls ') and @placeholder='Search']"},
         };
     }